+++ commerce_payzen/src/Controller/PaymentNotificationController.php
@@ -1,12 +1,12 @@
 <?php
 /**
- * Copyright (C) 2017-2018 Lyra Network.
- * This file is part of PayZen for Drupal Commerce.
- * See COPYING.md for license details.
+ * Copyright © Lyra Network.
+ * This file is part of PayZen for Drupal Commerce. See COPYING.md for license details.
  *
- * @author Lyra Network <contact@lyra-network.com>
- * @copyright 2017-2018 Lyra Network
- * @license http://www.gnu.org/licenses/gpl.html GNU General Public License (GPL v2)
+ * @package   Payzen
+ * @author    Lyra Network <contact@lyra-network.com>
+ * @copyright Lyra Network
+ * @license   http://www.gnu.org/licenses/gpl.html GNU General Public License (GPL v2)
  */
 namespace Drupal\commerce_payzen\Controller;