+++ CHANGELOG.md
@@ -1,2 +1,9 @@
+2.0.1, 2019-01-14:
+- Bug fix: siganture error when using special characters as ' and ".
+- Compatibility with Drupal Commerce 2.10 (fix error at the end of the payment).
+- [prodfaq] Fix notice about shifting the shop to production mode.
+- Added Spanish translation.
+- Fix new signature algorithm name (HMAC-SHA-256).
+
 2.0.0, 2018-06-04:
 - New mdoule compatible with Drupal Commerce 8.x-2.x.